Read xls and xlsx files.

Read xls and xlsx files.

While read_excel() auto detects the format from the file extension, read_xls() and read_xlsx() can be used to read files without extension.

read_excel(path, sheet = 1L, col_names = TRUE, col_types = NULL,
  na = "", skip = 0, guess_max = 1000)

read_xls(path, sheet = 1L, col_names = TRUE, col_types = NULL, na = "",
  skip = 0, guess_max = 1000)

read_xlsx(path, sheet = 1L, col_names = TRUE, col_types = NULL, na = "",
  skip = 0, guess_max = 1000)

Arguments

path

Path to the xls/xlsx file

sheet

Sheet to read. Either a string (the name of a sheet), or an integer (the position of the sheet). Defaults to the first sheet.

col_names

TRUE to use the first row as column names, FALSE to get default names, or a character vector giving a name for each column. If user provides col_types as a vector, col_names can have one entry per column, i.e. have the same length as col_types, or one entry per unskipped column.

col_types

Either NULL to guess from the spreadsheet or a character vector containing one entry per column from these options: "skip", "numeric", "date", "text" or "list". The content of a cell in a skipped column is never read and that column will not appear in the data frame output. A list cell loads a column as a list of length 1 vectors, which are typed using the type guessing logic from col_types = NULL, but on a cell-by-cell basis.

na

Character vector of strings to use for missing values. By default, readxl treats blank cells as missing data.

skip

Number of rows to skip before reading any data. Leading blank rows are automatically skipped.

guess_max

Maximum number of rows to use for guessing column types.

Examples

datasets <- readxl_example("datasets.xlsx")
read_excel(datasets)

#> # A tibble: 150 × 5
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#>          <dbl>       <dbl>        <dbl>       <dbl>   <chr>
#> 1          5.1         3.5          1.4         0.2  setosa
#> 2          4.9         3.0          1.4         0.2  setosa
#> 3          4.7         3.2          1.3         0.2  setosa
#> # ... with 147 more rows

# Specific sheet either by position or by name
read_excel(datasets, 2)

#> # A tibble: 32 × 11
#>     mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>   <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1  21.0     6   160   110  3.90 2.620 16.46     0     1     4     4
#> 2  21.0     6   160   110  3.90 2.875 17.02     0     1     4     4
#> 3  22.8     4   108    93  3.85 2.320 18.61     1     1     4     1
#> # ... with 29 more rows

read_excel(datasets, "mtcars")

#> # A tibble: 32 × 11
#>     mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>   <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1  21.0     6   160   110  3.90 2.620 16.46     0     1     4     4
#> 2  21.0     6   160   110  3.90 2.875 17.02     0     1     4     4
#> 3  22.8     4   108    93  3.85 2.320 18.61     1     1     4     1
#> # ... with 29 more rows

# Skipping rows and using default column names
read_excel(datasets, skip = 148, col_names = FALSE)

#> # A tibble: 3 × 5
#>    X__1  X__2  X__3  X__4      X__5
#>   <dbl> <dbl> <dbl> <dbl>     <chr>
#> 1   6.5   3.0   5.2   2.0 virginica
#> 2   6.2   3.4   5.4   2.3 virginica
#> 3   5.9   3.0   5.1   1.8 virginica

# "list" col_type can handle information of disparate types
df <- read_excel(readxl_example("clippy.xlsx"), col_types = c("text", "list"))
df

#> # A tibble: 4 × 2
#>                   name      value
#>                  <chr>     <list>
#> 1                 Name  <chr [1]>
#> 2              Species  <chr [1]>
#> 3 Approx date of death <dttm [1]>
#> # ... with 1 more rows

df$value

#> [[1]]
#> [1] "Clippy"
#> 
#> [[2]]
#> [1] "paperclip"
#> 
#> [[3]]
#> [1] "2007-01-01 UTC"
#> 
#> [[4]]
#> [1] 0.9
#>