readxl_progress {readxl}	R Documentation

Determine whether to show progress spinner

Description

By default, readxl displays a progress spinner unless one of the following is TRUE:

The spinner is explicitly disabled by setting options(readxl.show_progress = FALSE).

The code is run in a non-interactive session (interactive() is FALSE).

The code is run by knitr / rmarkdown.

The code is run in an RStudio notebook chunk. readxl uses the progress package under-the-hood and therefore is also sensitive to any options that it consults.

Usage

readxl_progress()

[Package readxl version 1.4.3 Index]