Determine file format
Source: R/excel-format.R

Determine if files are xlsx or xls. First the file extension is consulted. If that is unsuccessful and guess = TRUE and the file exists, the format is guessed from the file signature or "magic number".

excel_format(path, guess = TRUE)

Arguments

path
Path to the xls/xlsx file

guess
Logical. Whether to guess format based on the file itself, if the extension is neither "xlsx" nor "xls".

Value

Character vector with values "xlsx", "xls", or NA.

Examples

files <- c(
  "a.xlsx",
  "b.xls",
  "c.png",
  file.path(R.home("doc"), "html", "logo.jpg"),
  readxl_example("clippy.xlsx"),
  readxl_example("deaths.xls")
)
excel_format(files)

#> [1] "xlsx" "xls"  NA     NA     "xlsx" "xls"