read_excel {readxl}	R Documentation

Read xls and xlsx files.

Description

Read xls and xlsx files.

Usage

read_excel(path, sheet = 1, col_names = TRUE, col_types = NULL, na = "",
  skip = 0)

Arguments

path
Path to the xls/xlsx file

sheet
Sheet to read. Either a string (the name of a sheet), or an integer (the position of the sheet). Defaults to the first sheet.

col_names
Either TRUE to use the first row as column names, FALSE to number columns sequentially from X1 to Xn, or a character vector giving a name for each column.

col_types
Either NULL to guess from the spreadsheet or a character vector containing "blank", "numeric", "date" or "text".

na
Missing value. By default readxl converts blank cells to missing data. Set this value if you have used a sentinel value for missing values.

skip
Number of rows to skip before reading any data.

Examples

datasets <- system.file("extdata/datasets.xlsx", package = "readxl")
read_excel(datasets)

# Specific sheet either by position or by name
read_excel(datasets, 2)
read_excel(datasets, "mtcars")

[Package readxl version 0.1.1 Index]