readxl-package {readxl}	R Documentation

readxl: Read Excel Files

Description

Import excel files into R. Supports '.xls' via the embedded 'libxls' C library <https://github.com/libxls/libxls> and '.xlsx' via the embedded 'RapidXML' C++ library <http://rapidxml.sourceforge.net>. Works on Windows, Mac and Linux without external dependencies.

Author(s)

Maintainer: Jennifer Bryan jenny@rstudio.com (ORCID)

Authors:

Hadley Wickham hadley@rstudio.com (ORCID)

Other contributors:

RStudio (Copyright holder of all R code and all C/C++ code without explicit copyright attribution) [copyright holder, funder]

Marcin Kalicinski (Author of included RapidXML code) [contributor, copyright holder]

Komarov Valery (Author of included libxls code) [contributor, copyright holder]

Christophe Leitienne (Author of included libxls code) [contributor, copyright holder]

Bob Colbert (Author of included libxls code) [contributor, copyright holder]

David Hoerl (Author of included libxls code) [contributor, copyright holder]

Evan Miller (Author of included libxls code) [contributor, copyright holder]

See Also

Useful links:

https://readxl.tidyverse.org

https://github.com/tidyverse/readxl

Report bugs at https://github.com/tidyverse/readxl/issues

[Package readxl version 1.4.0 Index]